requirements.txt
Flask
python-dotenv
matplotlib
pandas
plotly
chardet
cchardet
openai
pdfplumber
PyMuPDF
PyPDF2
scikit-learn